334. Optimize array processing with two-pointer technique
Replace the O(n^3) brute-force approach with a two-pointer method to enhance performance. The previous used triple nested loops to evaluate all possible subarrays, this failed due to time complexity on execution on large data inputs.

The new approach uses leftMin and rightMax arrays to preprocess the data. One pointer starts from the beginning (for leftMin) and the other from the end (for rightMax) of the array.

Moved 334.IncreasingTripletSubsequence to Complete/LeetCode75/Array/String.